Reference

安装工具包DIPTutorials:

]add https://github.com/johnnychen94/Image-Processing-in-Julia

DIPTutorials.DIPTutorials — Module

A helper package for DIP tutorials: https://johnnychen94.github.io/Image-Processing-in-Julia/dev/

Examples

using DIPTutorials, FileIO
imgfile = DIP3e.get_filepath("0101") # figure 1.1
img = load(imgfile)

DIPTutorials.DIP3e.get_filepath — Method

get_filepath(img_name)

Return the path to the image file with img_name in its name.

Examples

using DIPTutorials
DIP3e.get_filepath("0101") # figure 1.1

See also: DIP3e.list_filenames

DIPTutorials.DIP3e.list_filenames — Method

list_filenames(idx::Integer; abspath=false)

Return a list of filenames available in chapter idx.

If abspath is true, then each item is a path that you can directly load.

Examples

using DIPTutorials
DIP3e.list_filenames(1) # all figures in chapter 1

See also: DIP3e.get_filepath